Change the weatherconditions arrray
Change from string replacement to a key=>value array and use the weather forecast code for the key - so the api code is our key for the conditionstext. with str_replace there where some weird issues and some codes from the api wasn't in the module. This API codes are listed on this page: https://developer.yahoo.com/weather/documentation.html#codes